Adds file creation timestamp header
For Ashur to be able to write a created-timestamp on CompositeFrames produced from a filtering run, it needs to receive a header containing the timestamp to write.

This commit ensures that before triggering filtering in Ashur, we set this created-timestamp on the exchange before triggering the filtering job in Ashur.