Adds file creation timestamp header (#729)
* Adds file creation timestamp header

For Ashur to be able to write a created-timestamp on CompositeFrames produced from a filtering run, it needs to receive a header containing the timestamp to write.

This commit ensures that before triggering filtering in Ashur, we set this created-timestamp on the exchange before triggering the filtering job in Ashur.

* Fixes timestamp precision issue in test

* Adds message to a MardukException being thrown

* Adds test for FileNameAndDigestIdempotentRepository

* Refactoring + adds tests

Moves logic for setting file created timestamp header to its own Processor to improve testability.
Adds test for mentioned Processor.

* Removes redundant log message